+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+
+- Allow empty environment variables for the directory
+
 ## v2.16.0 - 2023-07-24
 
 ### Fixed